Assistant checkpoint: Add supplier ingredient units functionality
Assistant generated file changes:
- client/src/services/unitService.ts: Add getSupplierIngredientUnits function
- client/src/components/documents/EditInvoiceSlider.tsx: Add supplier ingredient units query

---

User prompt:

OK now we want to get the ingredient units.
Basically, in the dropdown for the unit, we will split the sections in "already associated units" and "new units".

Since ingredients have already some units associated, we suggest them first.
For now let's not focus on the unit dropdown but just on getting the data.

For this we use our service unitService where we need to get the data.
We need to query the backend on : GET /units/v2/restaurant/{r_uuid}/supplier/{supplier_uuid}/ingredients
I've pasted an example of response :

{
  "data": [
    {
      "ingredient_uuid": "24d00317-ea18-441a-899b-517f8be72892",
      "units": [
        {
          "unit_name": "pack",
          "unit_uuid": "d577c9c8-0987-42ba-b470-b615b2fa8ef8"
        }
      ]
    },
    {
      "ingredient_uuid": "67a851e5-276b-4699-90dc-658a1197e35d",
      "units": null
    },
    {
      "ingredient_uuid": "caa3f201-2c1b-4163-bb80-fee308729b07",
      "units": [
        {
          "unit_name": "kg",
          "unit_uuid": "804ba27e-a7cb-4c39-8ae6-c79aabfab0f4"
        }
      ]
    }
  ],
  "message": "Success",
  "success": true
}

Can we build :
- The service function that get this data
- Get the data in editInvoiceSlider based on the selected supplier on the invoice and the restaurant