Assistant checkpoint: Fix preparation quantity initialization and field name
Assistant generated file changes:
- client/src/components/menu/ProductSheet.tsx: Set default quantity when adding preparation
- client/src/types/menu.ts: Fix typo in quantity field name

---

User prompt:

What happens is this :

hook.js:608 Warning: A component is changing an uncontrolled input to be controlled. This is likely caused by the value changing from undefined to a defined value, which should not happen. Decide between using a controlled or uncontrolled input element for the lifetime of the component. More info: https://reactjs.org/link/controlled-components Error Component Stack
    at input (<anonymous>)
    at _c (input.tsx:9:6)
    at form.tsx:107:9
    at div (<anonymous>)
    at form.tsx:76:6
    at FormField (form.tsx:33:6)
    at div (<anonymous>)
    at div (<anonymous>)
    at div (<anonymous>)
    at form (<anonymous>)
    at div (<anonymous>)
    at _c3 (sheet.tsx:57:6)
    at RecipeSheet (ProductSheet.tsx:193:3)
    at div (<anonymous>)
    at Products (Products.tsx:29:35)
    at div (<anonymous>)
    at div (<anonymous>)
    at Menu (Menu.tsx:8:45)
    at div (<anonymous>)
    at RestaurantProvider (RestaurantContext.tsx:14:38)
    at App (<anonymous>)

When I add a preparation, the default quantity is undefined
So I need to set the quantity before selecting the ingredient, can we fix this?