Assistant checkpoint: Create modifiers page with service functions and types
Assistant generated file changes:
- client/src/types/menu.ts: Add Modifier type definition
- client/src/services/menuService.ts: Add modifier service functions
- client/src/pages/menu/Modifiers.tsx: Create Modifiers page
- client/src/pages/menu/Menu.tsx: Update Menu component to show modifiers

---

User prompt:

Let's build the modifier page, this page will be highly similar to the product page in term of structure. Only difference is that instead of product_uuid, it's modifier_uuid and instead of product_name it's modifier_name.

Also, we will get the modifiers using the following endpoint:
/menu/v2/restaurant/{r_uuid}/modifiers

Can we build the page for modifiers, connect it to the corresponding menu section and add the correct service functions in the menu service