Assistant checkpoint: Fix document deletion and thumbnail update
Assistant generated file changes:
- client/src/components/documents/EditInvoiceSlider.tsx: Fix thumbnail deletion functionality

---

User prompt:

in @client/src/components/documents/EditInvoiceSlider.tsx it seems that the thrash to delete one of the images from the thumbnails does not work really well, when we click on the thrash nothing happens, we need to then click on a thumbnails to see the change and it looks bug.

Can we jst make sure that when a delete button is pressed, it deletes the thumbnails as well as the image from the list and updates the invoice?